Code for batch deletion (multithreading)
